PCQ-1782 taking value from functional.yaml

+++ src/functionalTest/java/uk/gov/hmcts/reform/pcqbackend/controllers/PcqRecordWithoutCaseTest.java
@@ -83,7 +83,7 @@ public class PcqRecordWithoutCaseTest extends PcqBaseFunctionalTest {
             answerRequest = jsonObjectFromString(jsonStringRequest);
             answerRequest.setPcqId(thirdUuid);
             answerRequest.setCompletedDate(PcqUtils.convertTimeStampToString(PcqUtils.getDateTimeInPast(
-                Long.parseLong(daysLimit) + 1)));
+                Integer.parseInt(daysLimit) + 1)));
 
             response = pcqBackEndServiceClient.createAnswersRecord(answerRequest);
